Add STEMlab discovery using Avahi
This is currently hidden behind a flag in the Makefile as it introduces
two dependencies, namely Avahi and libcurl.

The rationale for this addition is that the STEMlab (also known as Red
Pitaya, the name under which it was previously sold) does start SDR
applications installed on it automatically. To avoid having to open the
web interface in order to use it as an SDR (which requires knowing its
IP or hostname, provided hostname resolution works), this patch adds
functionality to pihpsdr to discover the STEMlab using Avahi, and to
start a selected SDR application using its web API.