Fix STEMlab discovery without hostname resolution
Before, the web request to start the SDR application was sent to the
STEMlab identified by the hostname (rp-xxxxxx). Unfortunately, there are
cases where the STEMlab can be found via IP address, but not via
hostname.

This should have really been in with the original commit for STEMlab
discovery already, but seems to have slipped under my radar - listing
the installed HPSDR applications already uses the IP address, but
actually starting it was still left with using hostname resolution.