client_server: fix off-by-one array usage in generate_pwd_hash()
array in C starts at 0 index therefore the last value is LENGTH-1.
It was working because some OSes memset() to zero the `s` array but this
is not guaranteed by C standard and in fact if you have garbage in `s`
the authentication will fail. Moreover, if the password is greater or
equal to SHA512_DIGEST_LENGTH a segmentation fault will arise since
array access is out of bound (64+64+5 > 2*64+4); see line 240.